+++ CHANGELOG.md
@@ -16,8 +16,7 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 
 ### Fixed
 
-- **[BC Break]** `Parameters::values` decode `ByteSequence` and `Token` classes.
-- **[BC Break]** `Parameters::values` no longer throw on invalid state; value is skipped from returned array.
+- None.
 
 ### Deprecated
 
@@ -34,6 +33,7 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 - **[BC Break]** `::sanitize` method is removed use `Parameters::clear` method instead if needed.
 - **[BC Break]** `isEmpty` method is removed use `hasMembers` method instead.
 - **[BC Break]** `Parameters::value` use `Item::value` method instead.
+- **[BC Break]** `Parameters::values` use `Parameters::getIterator` instead.
 
 ## [0.5.0] - 2022-05-13